+++ Keyrita/Measurements/Redirects.cs
@@ -37,7 +37,7 @@ namespace Keyrita.Measurements
         protected override void Compute()
         {
             SortedTrigramSetResult tgSet = (SortedTrigramSetResult)AnalysisGraphSystem.ResolvedNodes[eInputNodes.SortedTrigramSet];
-            long totalTgs = tgSet.TrigramCoverage;
+            double totalTgs = tgSet.TrigramCoverage;
 
             TrigramStatsResult tgs = (TrigramStatsResult)AnalysisGraphSystem.ResolvedNodes[eInputNodes.TrigramStats];
             mResult.TotalRedirects = tgs.TotalRedirects / totalTgs * 100;
